Rename `write_to_dataframe.py` to `write_to_data_table.py`.
