+++ metricflow-semantics/metricflow_semantics/dag/id_prefix.py
@@ -43,7 +43,7 @@ class StaticIdPrefix(IdPrefix, Enum, metaclass=EnumMetaClassHelper):
     DATAFLOW_NODE_PASS_FILTER_ELEMENTS_ID_PREFIX = "pfe"
     DATAFLOW_NODE_READ_SQL_SOURCE_ID_PREFIX = "rss"
     DATAFLOW_NODE_WHERE_CONSTRAINT_ID_PREFIX = "wcc"
-    DATAFLOW_NODE_WRITE_TO_RESULT_DATAFRAME_ID_PREFIX = "wrd"
+    DATAFLOW_NODE_WRITE_TO_RESULT_DATA_TABLE_ID_PREFIX = "wrd"
     DATAFLOW_NODE_WRITE_TO_RESULT_TABLE_ID_PREFIX = "wrt"
     DATAFLOW_NODE_COMBINE_AGGREGATED_OUTPUTS_ID_PREFIX = "cao"
     DATAFLOW_NODE_CONSTRAIN_TIME_RANGE_ID_PREFIX = "ctr"
